refactor: rename modules in imports
